Move and rename data view's calls to serviceFacade folder
