+++ src/serviceFacade/userServiceFacade.js
@@ -1,5 +1,18 @@
 import callApi from "../common/callApi";
 
+/**
+ * Get user information including name, institution, email, etc.
+ * @param userIds An array of user ids
+ * @returns {Promise<any>}
+ */
+export const getUserInfo = ({ userIds }) => {
+    const userQuery = userIds.join("&username=");
+
+    return callApi({
+        endpoint: `/api/user-info?username=${userQuery}`,
+    });
+};
+
 function getUserProfile(key) {
     return callApi({
         endpoint: "/api/profile",
